Create linearSearch.c
first input the number of elements in the array, then input the elements themselves. After that, the user inputs the element they want to search for. The linearSearch function is used to perform the search, and it returns the index of the element if found, or -1 if not found. The result is then printed to the user.